+++ docs/changelog.rst
@@ -1,9 +1,15 @@
 Change log
 ==========
 
-v0.0.20 <2019-xx-xx>
+v0.0.21 <2020-xx-xx>
+--------------------
+- `#95`_: Fixed parameter generation error of MLPG.
 
+
+v0.0.20 <2020-xx-xx>
+--------------------
 - `#90`_: Expand available speakers list for CMU Arctic dataset
+- `#91`_: Add support for JVS dataset: a free Japanese multi-speaker voice corpus.
 
 --------------------
 
@@ -186,3 +192,4 @@ v0.0.1 <2017-08-14>
 .. _#87: https://github.com/r9y9/nnmnkwii/pull/87
 .. _#88: https://github.com/r9y9/nnmnkwii/pull/88
 .. _#90: https://github.com/r9y9/nnmnkwii/pull/90
+.. _#91: https://github.com/r9y9/nnmnkwii/issues/91